Move createFullCreditsMethod to custom hook
